snow/stm32: re-configure power LED on the fly (input vs. pwm)
Usually the power LED is driven by the PWM mode so that its nominal
brightness can be set to a "soft" on value. However, when
the LED is to remain off the LED should be switched back to general
input mode and enable the pull-down resistor. This reduces
voltage leakage.

This CL updates the power_led_task to configure the LED however is
appropriate and adds board functions to re-configure the GPIO.

Signed-off-by: David Hendricks <dhendrix@chromium.org>

BUG=chrome-os-partner:12381
TEST=None.

Change-Id: If90ac78aaffe7358cce80dd02ec1423c2cb4f664