host_command: Add optional flags to host commands
If CONFIG_EXTRA_HC_FLAGS is defined, we add a flags byte to each command
to indicate whether the command executes immediately, or takes time
(e.g. flash erase takes time). Slow commands must be queued for later
processing, and an interim response sent back to the host.

NOTE: this feature is of no use for LPC, since all command execution is
synchronous.

BUG=chrome-os-partner:10533
TEST=manual
build and boot to kernel on snow

Change-Id: I5a97ef24260f3fb7b965366e24f82f1f161c435c
Signed-off-by: Simon Glass <sjg@chromium.org>