host_command: Add GET_STATUS command to check EC comms status
Some commands cause the EC to process for a while. On protocols other than
LPC this may be visible to the AP, in that new commands may not be accepted
until the old one is completed. Add a command to check the status of the EC,
to find out if it is processing the last command.

BUG=chrome-os-partner:10533
TEST=manual
build and boot to kernel on snow

Change-Id: I023fa89e66bd484605cacef4b455f8e3d69c67df
Signed-off-by: Simon Glass <sjg@chromium.org>