Replacing usleep with interrupt friendly udelay
There was a usleep put into a function that can be called from in an
interrupt context, which doesn't work.  This just switches it over to
udelay which will work in an interrupt.

Also flips the condition on the i2c if/else that might send it there.
It was backwards before

BUG=chrome-os-partner:12688
TEST=Run "battery" "pmu" boot the machine and use the keyboard.  Then
replace the in_interrupt_context() function with "1" to force it to use
polling and repeat the test.  Everything should work in both cases.

Change-Id: Ib2a8c7f9e5e2eb3f6b00678d6307afc9dd5f0518
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>