host_command: Add special handling for GET_STATUS
This command should be executed immediately. Instead of a comnand flag,
just add a special case for now. The rest of the logic is handled in
the driver.

Note: we could do this code in the driver also, but that would mean
the driver would call host_command_process() and host_send_response()
itself. ()I feel much of this could would be better in host_command.)

BUG=chrome-os-partner:12685
BRANCH=snow,link
TEST=manual
build and boot to kernel on snow
(link needs testing but my link doesn't boot)

Change-Id: I56458f7b9083d63622141711dee626f6209915c1
Signed-off-by: Simon Glass <sjg@chromium.org>