Update patch set 3
Patch Set 3:

Apologies for the delay on this... I had a chance to try out the previous patch some more and it turned out that making PMIC_PWRON_L depend on sysjump is a bad idea because, obviously, a normal AP shutdown will not cause the EC to shutdown and therefore the sysjump flag can remain set.

After a bit more thought, checking XPSHOLD seems like the most sensible option here... If the XPSHOLD is asserted (AP is on), then don't re-initiate the PMIC power on sequence. 

LMK what you guys think.

Patch-set: 3
Label: Code-Review=0