Snow: i2c arbitration cleanup (suspend)
The ec used to have to check the state of the AP since it would leave
the arbitration lines when it suspended.  That meant you couldn't trust
the arbitration lines without checking the power state.  Now the AP
pulls the arbitration lines high when it suspends, so it no longer needs
to test this when trying to acquire master.

BUG=chrome-os-partner:12460
TEST=First, on the EC console run "battery" and "pmu" to make sure they
work.  Then repeat those steps after putting the AP in suspend by
running powerd_suspend, they should still work.  Then shutdown the
machine entirely and try them again.  Note: pmu needs the AP to be on to
work, so if it fails here that's okay, just make sure it's not an
arbitration error.
BRANCH=snow

Change-Id: I335156bbce4888949111f74e8a83fe9d184a7a63
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>