Update patch set 2
Patch Set 2:

I'm pretty sure the AP arbitration pins are only pulled high on suspend, not power-off, so that check is still necessary.  When it doesn't check for power-off, the EC can't grab the bus when the AP is off, so it can't talk to the battery, etc.

Added in change to the release function, so it still releases the lock when the AP is off, per: 12918

Patch-set: 2
Label: Code-Review=0