stm32: Add EC_FLASH_PROTECT_ENTIRE_NOW flag.
Current EC_FLASH_PROTECT_RW_NOW on stm32 is used to lock the entire
flash. This could lead confusion in the future. So, add a flag to do that.

However, u-boot is still using RW_NOW. To make the u-boot can run on both
old and new EC, we keep RW_NOW to support entire-chip protection.

Also fixed the nit that EC_FLASH_PROTECT_RW_AT_BOOT is actually not
supported on stm32.

BUG=chrome-os-partner:12951
BRANCH=snow
TEST=build in chroot only.
Signed-off-by: Louis Yung-Chieh Lo <yjlou@chromium.org>

Change-Id: I2395e93793f590e6fb8aae7006eb8e5c836002bc