Update patch set 2
Patch Set 2:

It sounds like this is just the way it's going to have to be.  The rail that the pull-ups are connected to gets powered down when the AP is off, so we just have to ignore the signal from the looks of it.

Here was Frank's response:
"The pullup resistors get pulled up to the CPU rail, which goes low when the CPU turns off.  These signals were originally used for SPI, so we didn't care about them when the CPU wasn't on."

"Considering we already gerbered out the MP version of the board, I think we're stuck.  Hopefully we can ignore these signals when the 3.3V rail is not enabled (EC controls this)..."

Patch-set: 2