Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

It seems to me that this is a symptom of the problem in the power button code:

http://gerrit.chromium.org/gerrit/4486

It is checking for the power button being held, but should only check for a press as the comment says.

Rather than contort the logic in the EC I think we could just detect a press in the U-Boot code. I will take a look.

Re USB, we hold power down with servo until SPL has started, so there is no particular requirement there now. There used to be...

Patch-set: 1
Reviewer: Gerrit User 1000294 <1000294@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1
Label: Verified=0