snow: enable pull-up on PA7 (SUSPEND_L)
This changes the configuration of SUSPEND_L from a floating input
to an input with pull-up enabled. This results in much quicker power
LED response when the AP comes out of suspend to ensure that the LED
stops "breathing" before other components (like the display) come up.

Signed-off-by: David Hendricks <dhendrix@chromium.org>
BRANCH=snow
BUG=chrome-os-partner:12700
TEST=Tested using powerd_suspend on Snow, waiting for breathing
effect to begin, then striking a key and seeing the LED turn
solid immediately.

Change-Id: I62f02324a2a1fbfb6eff539fc6fdc35a035fa020