snow: enable pull-up on PA7 (SUSPEND_L) when machine is suspended
This changes the configuration of SUSPEND_L from a floating input
to an input with pull-up enabled when the machine goes into suspend
mode. This results in much quicker power LED response when the AP
comes out of suspend to ensure that the LED stops "breathing" before
other components (like the display) come up.

This is only needed when we're waiting for a suspend --> resume
transition. So to avoid unncessary current leakage we'll configure
SUSPEND_L as a floating input for other states (EC init, AP resume,
AP shutdown).

Signed-off-by: David Hendricks <dhendrix@chromium.org>
BRANCH=snow
BUG=chrome-os-partner:12700
TEST=see below

Test procedure
A. Visual inspection:
1. Run powerd_suspend on Snow, wait for breathing effect to begin
2. Strike a key and seeing the LED turn solid immediately.

B. Walk thru power states and dump register from EC console:
(note: bits 31:28 are PA7 state, 4 for floating, 8 for pull-up)
1. Machine halted, EC rebooted, read 0x40010800 = 0x41184144
2. Press power button, read 0x40010800 = 0x41184144
3. Login and run powerd_suspend, read 0x40010800 = 0x81184144
4. Press a key to resume, read 0x40010800 = 0x41184144
5. Shutdown, read 0x40010800 = 0x41184144

Change-Id: I62f02324a2a1fbfb6eff539fc6fdc35a035fa020