Update patch set 1
Patch Set 1:

Simon,
The new EC code is actually simplifying the EC state machine.  It basically switches the order in which we check xpshold and power release.  It used to be power release and then xpshold and now its the opposite.  The fact that uboot has also fixed this problem is great.  I am going to address Vincent's and Dave's comments and then proceed with the checkin.

Patch-set: 1