Update patch set 3
Patch Set 3:

I'm not sure about your modification :
first of all, we probably only really care of the leakage in low power modes (suspend and off) and it seems we still the pull up in suspend since it's open drain.
In run mode, SUSPEND_L is high, we won't leak through the pull up, so we might not need to make the input floating (and I'm a  bit worried that toggling input type at the same time it's transitioning in suspend/resume might introduce some unforesen glitch / oscillation)
In off mode, we might leak in that case. 
I

Patch-set: 3
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0