stm32f: only change trigger selection if needed
(WIP: this is just uploaded so it doesn't get lost)

This adds trivial checks to read the rising/falling trigger selection
registers to determine if they actually need to be written.

Though on issues have been noticed yet, this is intended to avoid
the corner case of missing an edge while writing the register.

BRANCH=none
BUG=None
TEST=not yet...
Signed-off-by: David Hendricks <dhendrix@chromium.org>

Change-Id: I9558f6eb342e014fc6d60aebb8224c316767ed9b