stm32f: remove gpio_set_alternate_function()
The gpio_set_alternate_function() functino for STM32F is not used,
and even if it were it would be incorrect. So for now it just
takes up space (and doesn't even look good doing so).

Unlike STM32L, alternate functions rely on toggling bits in AFIO
remapping registers rather than setting a simple AF number. This
would make writing a working version of this function trickier,
and it may not be worth the effort.

Signed-off-by: David Hendricks <dhendrix@chromum.org>
BRANCH=snow
BUG=none
TEST=locally compiled for snow

Change-Id: I2ce1e7aba2760a94819500af4e322812f3346ad3