Update patch set 2
Patch Set 2: Verified

Note that I think this bug also could affect things in the RO-RW transition, which I think happens during flashrom.  ...so it could be affecting real behavior.  I'll file a bug and add a link to it here along with scope shots.

Let's get this in and in the firmeware branch.

Patch-set: 2
Label: Verified=+1