Update patch set 2
Patch Set 2:

This is data which changes across code versions, and must not get zero-init'd or overwritten by init.S as the RO/RW image sizes change.  So sticking it at the top of RAM is convenient.

Unfortunately, Snow is so tight on RAM that there is no free RAM at the top.

I'll look into how we can top-justify this data in the linker.

Note that the save data was previously statically allocated on Snow, so this change alone shouldn't make anything worse than it was.

Patch-set: 2