Calibrate IR temperature sensors for Link DVT
The TMP006 IR sensors are calibrated against the temperature measured on
the *external* side on the casing using a thermocouple stick at the
vertical of the sensor.

The hinge sensor is sending back strange values, and the Tobject from
there should not be trusted.

The DC-Jack C-case sensor is not calibrated (and will be removed soon).

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=None
TEST=on Link DVT, compare EC temperature values against thermocouple
readings.
BRANCH=link

Change-Id: I03375dd1c2f3a0aa56b0d2f343dad3b8f7581bc2