Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(1 inline comment)

One last thing.  Note that a lot of these cases are pretty exceptional and hard to test, so we're just making our best guess.  It think that's fine.  It might be worth it to print something to the logs in case these very exceptional cases happen so we don't confuse ourselves further.

Patch-set: 2
Label: Commit-Queue=0
Label: Verified=0