Update patch set 2
Patch Set 2: Looks good to me, approved

(3 inline comments)

Reset was the only thing that was required and Charlie has pointed out that it's already doing it OK.  Adding logging is optional, but would be nice.

Note that I might spend 30 minutes trying to see if I can get U-Boot to bit-bang a "clock stretch" a transaction to TPSChrome to test this--we'll see how things are going today.

Patch-set: 2
Label: Code-Review=+2