Update patch set 4
Patch Set 4:

In some final re-testing, it turns out that just setting that software reset bit on the i2c peripheral isn't good enough.  It seems that it resets the whole config, so I switched back to using i2c_init's to reset the peripheral.  However, I pulled the lines that enable the interrupts out and moved them to the i2c_init call, that usually wraps the i2c_init1 and i2c_init2 calls every other time they're called.  This way it should reset the peripheral entirely but not reset the irqs in this case, but leaves the initial configs the same.

Patch-set: 4
Label: Code-Review=0
Label: Commit-Queue=0
Label: Verified=0