Update patch set 6
Patch Set 6: (6 inline comments)

Filed http://crosbug.com/p/13742 to address djkurtz feedback (thanks for review!)

Nothing here seems showstopper, so filed it as P2.  Doing the bitbanging tests seems like it would be slightly more urgent so might make sense to tackle earlier.  ...but I don't think I can afford to spend any more time on this at the moment and I'm pretty convinced that we're at least solid enough for all the peripherals that are on our bus.  ...and if we do have some failure case it _should_ be caught by the later PMIC reset.

If someone has time to pick this up that'd be great.  Otherwise I'll try to jump back to it when time permits.

Thanks!

Patch-set: 6