Snow: Always reset i2c when it's initiallized
Previously, the i2c init code would only preform a software reset of the
i2c peripheral it is initializing when it was already BUSY.  It turns
out it's always BUSY and the init functions are now used in two other
places where they always want the software reset as well, so this pulls
out the conditional, and makes it always do it.

BUG=chrome-os-partner:13388
TEST=Standard i2c stress tests.  Running a loop of i2cdumps from the AP
while looping i2c transactions on the EC run without any errors.  Even
across multiple reboots, and jumping back and forth from RO to RW on the
EC via sysjump while the AP is still stressing the bus.
BRANCH=snow

Change-Id: I6b3aaae0042844033bb04cf5cb4171c8be041ad9
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>