Check boot key combos if refresh key is held down
Instead of how it is now, where the boot key combinations are only
tested if it was a keyboard-controlled reset.  This is important for
testing/debugging EC software sync, which has a tendancy to blow away
your RW EC as soon as you flash a test EC and it reboots.  Now you can
hold down refresh+downarrow while flashing.

This does not affect keyboard-controlled dev switching, since that's
done in the AP after the EC boots.  It also does not add any new key
combos, just makes it possible to trigger the existing ones without a
Silego reset.

BUG=chrome-os-partner:13753
BRANCH=link
TEST=manual

1. Boot normally.  Works.
2. Power+Refresh.  Boots normally.
3. Power+Refresh+Esc.  Boots to recovery.
4. Power+Refresh+Down.  EC reboots, system powers down.
5. Hold down Esc and reboot from EC console.  Boots normally.
6. Hold down Refresh+Esc and reboot from EC console.  Boots to recovery.

Change-Id: Iabe4fd13589428a40b83f591ea679cbc6f83959d
Signed-off-by: Randall Spangler <rspangler@chromium.org>