Add synthesized keyboard events for testing.
This defines a set of commands the AP can use to download and execute
a simple keystroke emulation program.  Because the EC has serious
memory limitations, the program is stored as bytecodes.  The bytecodes
also specify the delays between key events.

We want to be able to specify a sequence of events, rather than one
event at a time, so that their synthesis is more reliable.

BUG=chromium-os:31352
TEST=none for now
BRANCH=none

Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

Change-Id: I6c38a180155f49ca665f4b33e740b3a3d23113c7