Add synthesized keyboard events for testing.
This defines a set of commands the AP can use to download and execute
a simple keystroke emulation program.  Because the EC has serious
memory limitations, the program is stored as bytecodes.  The bytecodes
also specify the delays between key events.  This results in more precise
timing of keyboard events than if we requested individual events from the
kernel.

Also extend ectool to send keystroke emulation programs to the ec,
and add the kbprog tool to translate from a high-level notation
to bytecodes.

BUG=chromium-os:31352
TEST=manual
BRANCH=none

Signed-off-by: Luigi Semenzato <semenzato@chromium.org>

Change-Id: I6c38a180155f49ca665f4b33e740b3a3d23113c7