Update patch set 1
Patch Set 1:

Very little runtime, of course.

Right now I think we're at more risk of overwriting global variables in sysjump/panic.  That's not a problem as long as we're sitting in RO firmware, but once we want to jump to RW it's more of an issue.

Feel free to suggest less-aggressive tweaks to stack sizes.

Patch-set: 1