Update patch set 2
Patch Set 2:

Yes it seems to produce smaller and faster code, by avoiding some literal pool entries, in the case where we ave several functions that access the peripheral and are either not merged, or have register spills.

I am going to tidy up the SPI driver to use this method exclusively though, as at the moment it is using both!

Patch-set: 2