Snow: Don't check i2c arbitration if AP suspended
Recent changes have made it so that the i2c arbitration lines float when
the AP is in suspend to reduce power consumption.  Unfortunately,
without the pull-up resistors, this means that their values are not
reliable any more, so we have to check to see if the AP is in suspend
first before trusting those values.  If the AP is suspended, then we
automatically get the bus.

BUG=chrome-os-partner:13880
TEST=boot machine, unplug from AC power and close the lid to put it into
suspend.  From the EC console, make sure that it goes into suspend by
running "power" and confirm that there are no I2C arbitration errors
reported on the console.  Finally, run "pmu" to make extra sure the i2c
bus is still working.
BRANCH=snow

Change-Id: I90fc249707c18ba8d538c6292fedd7c145322c73
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>