Update patch set 2
Patch Set 2:

agreed -- let's make sure we're all in agreement here and not overlooking any details.

The change I submitted only re-configures the arbitration GPIOs on full power off (XPSHOLD goes to 0) and cold boot (XPSHOLD goes from 0 -> 1). So for CHIPSET_STATE_ANY_OFF EC_CLAIM was reconfigured as an input and it did not make sense to try to assert it.

This change adds a check for CHIPSET_STATE_SUSPEND (XPSHOLD remains 1, SUSPEND_L goes to 0). In this state, the GPIOs are still enabled with EC_CLAIM as an output. So it should be asserted.

Patch-set: 2